requirements.txt
pandas
pyautoguiL  REGA000441  Garmin